+++ AzureKeyVault/AzureClient.cs
@@ -1,16 +1,9 @@
 ﻿// Copyright 2023 Keyfactor
-// 
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// 
-//     http://www.apache.org/licenses/LICENSE-2.0
-// 
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
+// and limitations under the License.
 
 using System;
 using System.Collections.Generic;
@@ -35,7 +28,10 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
     public class AzureClient
     {
         internal protected virtual AkvProperties VaultProperties { get; set; }
+        private string ActiveTenantId { get; set; } // for searching across multiple tenants.
+
         ILogger logger { get; set; }
+
 
         private protected virtual CertificateClient CertClient
         {
@@ -43,6 +39,7 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
             {
                 if (_certClient != null)
                 {
+                    logger.LogTrace("getting previously initialized certificate client");
                     return _certClient;
                 }
 
@@ -52,17 +49,22 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
 
                 if (this.VaultProperties.UseAzureManagedIdentity)
                 {
+                    logger.LogTrace("Entering the managed identity workflow");
+
                     var credentialOptions = new DefaultAzureCredentialOptions();
 
                     if (!string.IsNullOrEmpty(this.VaultProperties.ClientId)) // are they using a user assigned identity instead of a system assigned one (default)?
                     {
+                        logger.LogTrace("they provided client ID, so it is a user assigned managed identity (instead of system assigned)");
                         credentialOptions.ManagedIdentityClientId = VaultProperties.ClientId;
                     }
                     cred = new DefaultAzureCredential(credentialOptions);
                 }
                 else
                 {
+                    logger.LogTrace("They are using a service principal to authenticate, generating the credentials");
                     cred = new ClientSecretCredential(VaultProperties.TenantId, VaultProperties.ClientId, VaultProperties.ClientSecret);
+                    logger.LogTrace("generated credentials", cred);
                 }
 
                 _certClient = new CertificateClient(new Uri(VaultProperties.VaultURL), credential: cred);
@@ -72,39 +74,74 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
         }
         protected CertificateClient _certClient { get; set; }
 
+        internal protected virtual ArmClient getArmClient(string tenantId)
+        {
+            TokenCredential credential;
+
+            if (this.VaultProperties.UseAzureManagedIdentity)
+            {
+                logger.LogTrace("getting management client for a managed identity");
+                var credentialOptions = new DefaultAzureCredentialOptions();
+                if (!string.IsNullOrEmpty(tenantId)) credentialOptions.TenantId = tenantId;
+
+                if (!string.IsNullOrEmpty(this.VaultProperties.ClientId)) // they have selected a managed identity and provided a client ID, so it is a user assigned identity
+                {
+                    logger.LogTrace("It is a user assigned managed identity");
+                    credentialOptions.ManagedIdentityClientId = VaultProperties.ClientId;
+                }
+                credential = new DefaultAzureCredential(credentialOptions);
+            }
+            else
+            {
+                logger.LogTrace("getting credentials for a service principal identity");
+                credential = new ClientSecretCredential(tenantId, VaultProperties.ClientId, VaultProperties.ClientSecret);
+                logger.LogTrace("got credentials for service principal identity", credential);
+            }
+
+            _mgmtClient = new ArmClient(credential);
+            logger.LogTrace("created management client", _mgmtClient);
+            return _mgmtClient;
+        }
         internal protected virtual ArmClient KvManagementClient
         {
             get
             {
                 if (_mgmtClient != null)
                 {
+                    logger.LogTrace("getting previously initialized management client");
                     return _mgmtClient;
                 }
+                return getArmClient(VaultProperties.TenantId);
 
                 // var subId = VaultProperties.SubscriptionId ?? VaultProperties.StorePath.Split("/")[2];
                 // var creds = SdkContext.AzureCredentialsFactory.FromServicePrincipal(VaultProperties.ApplicationId, VaultProperties.ClientSecret, VaultProperties.TenantId, AzureEnvironment.AzureGlobalCloud);
                 //NOTE: creating a certificate store from the platform is currently only supported for Azure GlobalCloud customers.
 
-                TokenCredential credential;
+                //TokenCredential credential;
 
-                if (this.VaultProperties.UseAzureManagedIdentity)
-                {
-                    var credentialOptions = new DefaultAzureCredentialOptions();
+                //if (this.VaultProperties.UseAzureManagedIdentity)
+                //{
+                //    logger.LogTrace("getting management client for a managed identity");
+                //    var credentialOptions = new DefaultAzureCredentialOptions();
 
 
-                    if (!string.IsNullOrEmpty(this.VaultProperties.ClientId)) // they have selected a managed identity and provided a client ID, so it is a user assigned identity
-                    {
-                        credentialOptions.ManagedIdentityClientId = VaultProperties.ClientId;
-                    }
-                    credential = new DefaultAzureCredential(credentialOptions);
-                }
-                else
-                {
-                    credential = new ClientSecretCredential(VaultProperties.TenantId, VaultProperties.ClientId, VaultProperties.ClientSecret);
-                }
+                //    if (!string.IsNullOrEmpty(this.VaultProperties.ClientId)) // they have selected a managed identity and provided a client ID, so it is a user assigned identity
+                //    {
+                //        logger.LogTrace("It is a user assigned managed identity");
+                //        credentialOptions.ManagedIdentityClientId = VaultProperties.ClientId;
+                //    }
+                //    credential = new DefaultAzureCredential(credentialOptions);
+                //}
+                //else
+                //{
+                //    logger.LogTrace("getting credentials for a service principal identity");
+                //    credential = new ClientSecretCredential(ActiveTenantId, VaultProperties.ClientId, VaultProperties.ClientSecret);
+                //    logger.LogTrace("got credentials for service principal identity", credential);
+                //}
 
-                _mgmtClient = new ArmClient(credential);
-                return _mgmtClient;
+                //_mgmtClient = new ArmClient(credential);
+                //logger.LogTrace("created management client", _mgmtClient);
+                //return _mgmtClient;
             }
         }
         protected virtual ArmClient _mgmtClient { get; set; }
@@ -113,12 +150,13 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
         public AzureClient(AkvProperties props)
         {
             VaultProperties = props;
-
+            ActiveTenantId = props.TenantId;
             logger = LogHandler.GetClassLogger<AzureClient>();
         }
 
         public virtual async Task<DeleteCertificateOperation> DeleteCertificateAsync(string certName)
         {
+            logger.LogTrace("calling method to delete certificate");
             return await CertClient.StartDeleteCertificateAsync(certName);
         }
 
@@ -134,10 +172,7 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
 
                 var vaults = resourceGroup.GetKeyVaults();
 
-
-                //TODO: Create store type parameter for Azure Location.
-
-                var loc = new AzureLocation(VaultProperties.VaultRegion); // pass property instead of hardcoded value after testing
+                var loc = new AzureLocation(VaultProperties.VaultRegion);
 
                 var skuType = VaultProperties.PremiumSKU ? KeyVaultSkuName.Premium : KeyVaultSkuName.Standard;
 
@@ -159,15 +194,20 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
         {
             try
             {
+                logger.LogTrace("checking to see if the certificate exists and has been deleted");
+
                 if (CertClient.GetDeletedCertificates().FirstOrDefault(i => i.Name == certName) != null)
                 {
-                    RecoverDeletedCertificateOperation recovery = await CertClient.StartRecoverDeletedCertificateAsync(certName);
+                    logger.LogTrace("certificate to import has been previously deleted, starting recovery operation.");
+                    var recovery = await CertClient.StartRecoverDeletedCertificateAsync(certName);
                     recovery.WaitForCompletion();
                 }
 
+                logger.LogTrace("begin creating x509 certificate from contents.");
                 var bytes = Convert.FromBase64String(contents);
                 var x509 = new X509Certificate2(bytes, pfxPassword, X509KeyStorageFlags.Exportable);
                 var certWithKey = x509.Export(X509ContentType.Pkcs12);
+                logger.LogTrace($"importing created x509 certificate named {1}", certName);
                 var cert = await CertClient.ImportCertificateAsync(new ImportCertificateOptions(certName, certWithKey));
                 return cert;
             }
@@ -184,7 +224,9 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
             Pageable<CertificateProperties> inventory = null;
             try
             {
+                logger.LogTrace("calling GetPropertiesOfCertificates() on the Certificate Client", CertClient);
                 inventory = CertClient.GetPropertiesOfCertificates();
+                logger.LogTrace("got a response", inventory);
                 var certQuantity = inventory.Count();
             }
             catch (Exception ex)
@@ -193,9 +235,13 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
                 throw;
             }
 
+            logger.LogTrace("retrieving each certificate from the response");
             foreach (var certificate in inventory)
             {
+                logger.LogTrace("getting details for the individual certificate", certificate);
                 var cert = await CertClient.GetCertificateAsync(certificate.Name);
+                logger.LogTrace("got certificate response", cert);
+
                 inventoryItems.Add(new CurrentInventoryItem()
                 {
                     Alias = cert.Value.Name,
@@ -208,21 +254,34 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
             return inventoryItems;
         }
 
-        public virtual async Task<List<string>> GetVaults()
+        public virtual List<string> GetVaults()
         {
+            var vaultNames = new List<string>();
             try
             {
-                SubscriptionResource subscription = await KvManagementClient.GetDefaultSubscriptionAsync();
-                ResourceGroupCollection resourceGroups = subscription.GetResourceGroups();
-                //ResourceGroupResource resourceGroup = await resourceGroups.GetAsync(this.VaultProperties.ResourceGroupName);
-                var vaultNames = new List<string>();
-
-                resourceGroups.ToList().ForEach(rg =>
+                if (VaultProperties.TenantIdsForDiscovery == null || VaultProperties.TenantIdsForDiscovery.Count() < 1)
                 {
-                    var rgVaults = rg.GetKeyVaults().ToList();
-                    vaultNames.AddRange(rgVaults.Select(v => v.Id.ToString()));
+                    throw new Exception("no tenant ID's provided.");
+                }
+                VaultProperties.TenantIdsForDiscovery.ForEach(tenantId =>
+                {
+                    logger.LogTrace($"getting ARM client for tenantId {tenantId}");
 
+                    var mgmtClient = getArmClient(tenantId);
+
+                    logger.LogTrace($"getting all available subscriptions in tenant with ID {tenantId}");
+                    var allSubs = mgmtClient.GetSubscriptions();
+
+                    logger.LogTrace($"got {allSubs.Count()} subscriptions");
+
+                    foreach (var sub in allSubs) {
+                        logger.LogTrace($"searching for vaults in subscription with ID {sub.Data.SubscriptionId}");
+                        var vaults = sub.GetKeyVaults();
+                        logger.LogTrace($"found {vaults.Count()} vaults.");
+                        vaultNames.AddRange(vaults.Select(v => v.Id.ToString()));
+                    }                    
                 });
+
                 return vaultNames;
             }
             catch (Exception ex)