+++ AzureKeyVault/Jobs/Discovery.cs
@@ -1,16 +1,9 @@
 ﻿// Copyright 2023 Keyfactor
-// 
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// 
-//     http://www.apache.org/licenses/LICENSE-2.0
-// 
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
+// and limitations under the License.
 
 using System;
 using System.Collections.Generic;
@@ -40,7 +33,7 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
             {
                 if (VaultProperties.TenantId == null) throw new MissingFieldException("Need to set Tenant Id value in directories to search field for discovery jobs that use a system managed identity.");
 
-                keyVaults = AzClient.GetVaults().Result;
+                keyVaults = AzClient.GetVaults();
             }
             catch (Exception ex)
             {