+++ CHANGELOG.md
@@ -1,9 +1,14 @@
-﻿- 2.0.1
+- 2.1
+  - Updated the Discovery job to support multiple tenants and all accessible subscriptions they contain
+  - Added more detailed trace logging during the discovery process
+
+- 2.0.1
   - Fixed bug when trying to parse Vault plan (premium/standard)
 
 - 2.0
   - Added support for Azure Managed Identity authentication
   - Updated Azure client libraries
   - Removed ObjectId parameter from StoreType definition
-  - Added SkuType and VaultRegion parameters to support vault creation from the platform.
+  - Added SkuType and VaultRegion parameters to support vault creation from the platform
 
+- 1.05